prod.requirements.txt
numpy==1.19.1
pandas==1.1.0
pickle-mixin==1.0.2
python-dotenv==0.14.0
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
urllib3==1.25.10
